+++ compiler/passes/docs.h
@@ -57,8 +57,8 @@ static inline std::string ltrim(std::string s);
 
 static inline bool isEmpty(std::string s);
 
-static std::string erase(std::string s, size_t count);
+static std::string erase(std::string s, int count);
 
-static size_t minimumPrefix(std::string s);
+static int minimumPrefix(std::string s);
 
 void ltrimAndPrintLines(std::string s, std::ofstream *file);
